Move some code from db-package to routers.
